fized sizing for wms who add borders internally
Works for xmonad, at least. It was not clear why XMoveResizeWindow()
would actually make the window bigger than requested(requested 604,
window ends up at 608 because borders are added on top of it).

If you use this and get different results in the sizing when your WM
tries to add borders, tweak the call to XMoveResizeWindow().