changed cursor drawing to an appended underscore
Not only was the rectangle for the cursor really ugly, it wouldn't line
up with some fonts(i.e. Japanese glyphs) because of how dmenu was
computing the width(only the first font in the font set was being used).

This changes it by making the cursor an appended underscore that looks a
bit better.