redone the cursor using drw_fontset_getwidth()
This time a buffer will store the next UTF-8 character(if present), and
set the width of the cursor to the width of that character. If we're at
the end of the line, instead the cursor will use the width of the
underscore('_') character.

Moreover, now that we catch FocusIn/FocusOut events, the cursor will
change when the window is/isn't focused, making it filled when it is(and
we're at the end of the line) and making it not when it isn't(or we're
on top of an existing character).