added actual support for multiple fonts
Unless I'm stupid and the documentation for Xft, fontconfig, etc. does
tell me how do specify a font set through a single XftFontOpenName()
call, dmenu was relying on fallbacks instead of actually taking a
fontset. This makes the fonts array sized at BUFSIZ and adds an option
to override defaults with any number of fonts up to that.