raising when obscured is undesirable when managed
That looks like a cryptic commit message, but what it means is that when
override_redirect is not set, it is undesirable for dmenu to raise its
window when it is obscured. The window manager that is taking care of
dmenu should make that decision on its own.

The XRaiseWindow() call is kept for when override_redirect is set and
still the window is obscured.