small refactor part 01
Who needs a config header file? Since dmenu encourages you to modify the
source code to configure it, like many other suckless projects, you're
kind of expected to know C, so looking at the compile units should be
good enough. I understand it'd be easier to have it all in a single
file, but this is the dmenu that *I* will be using, so...

I removed everything Xinerama, hopefully. I don't use it, so I don't
need it.