small refactor part 03 and how not to do cmd args
Moving variables around, still thinking about comments, added vim
modeline to the end of the file, and changed how command line args are
treated.

Optimal solution would've been to use getopt or getopt_long, but I went
with a switch and a hashing function to replace the if-else ladder. This
way hopefully the compiler creates a jump table and all. It was probably
already doing that, but refer to the previous commit aobut this being a
sickness.